+++ src/viadot/orchestration/prefect/tasks/__init__.py
@@ -9,15 +9,13 @@ from .duckdb import duckdb_query  # noqa: F401
 from .exchange_rates import exchange_rates_to_df  # noqa: F401
 from .git import clone_repo  # noqa: F401
 from .luma import luma_ingest_task  # noqa: F401
+from .mindful import mindful_to_df
 from .minio import df_to_minio  # noqa: F401
 from .redshift_spectrum import df_to_redshift_spectrum  # noqa: F401
 from .s3 import s3_upload_file  # noqa: F401
 from .sap_rfc import sap_rfc_to_df  # noqa: F401
-from .sharepoint import (
-    get_endpoint_type_from_url,  # noqa: F401
-    scan_sharepoint_folder,  # noqa: F401
-    sharepoint_download_file,  # noqa: F401
-    sharepoint_to_df,  # noqa: F401
-    validate_and_reorder_dfs_columns,  # noqa: F401
-)
-from .sql_server import sql_server_query, sql_server_to_df, create_sql_server_table
+from .sharepoint import get_endpoint_type_from_url  # noqa: F401
+from .sharepoint import scan_sharepoint_folder  # noqa: F401
+from .sharepoint import sharepoint_download_file  # noqa: F401
+from .sharepoint import sharepoint_to_df  # noqa: F401
+from .sharepoint import validate_and_reorder_dfs_columns  # noqa: F401