+++ src/viadot/sources/__init__.py
@@ -3,15 +3,17 @@ from importlib.util import find_spec
 from .cloud_for_customers import CloudForCustomers
 from .exchange_rates import ExchangeRates
 from .genesys import Genesys
+from .mindful import Mindful
 from .sharepoint import Sharepoint
-from .trino_source import Trino
 from .sql_server import SQLServer
+from .trino_source import Trino
 from .duckdb import DuckDB
 
 __all__ = [
     "CloudForCustomers",
     "ExchangeRates",
     "Genesys",
+    "Mindful",
     "Sharepoint",
     "Trino",
     "SQLServer",